`sd-indexer` crate
