+++ core/src/location/indexer/old_shallow.rs
@@ -15,6 +15,7 @@ use sd_file_path_helper::{
 	check_file_path_exists, ensure_sub_path_is_directory, ensure_sub_path_is_in_location,
 	IsolatedFilePathData,
 };
+use sd_indexer::rules::IndexerRule;
 use sd_utils::db::maybe_missing;
 
 use std::{
@@ -29,8 +30,7 @@ use tracing::{debug, error};
 
 use super::{
 	execute_indexer_save_step, iso_file_path_factory, location_with_indexer_rules,
-	old_walk::walk_single_dir, remove_non_existing_file_paths, rules::IndexerRule, IndexerError,
-	OldIndexerJobSaveStep,
+	old_walk::walk_single_dir, remove_non_existing_file_paths, IndexerError, OldIndexerJobSaveStep,
 };
 
 /// BATCH_SIZE is the number of files to index at each step, writing the chunk of files metadata in the database.
