+++ core/src/location/non_indexed.rs
@@ -12,6 +12,7 @@ use futures::Stream;
 use itertools::Either;
 use sd_file_ext::{extensions::Extension, kind::ObjectKind};
 use sd_file_path_helper::{path_is_hidden, MetadataExt};
+use sd_indexer::rules::{IndexerRule, RuleKind};
 use sd_prisma::prisma::location;
 use sd_utils::{chain_optional_iter, error::FileIOError};
 
@@ -32,10 +33,7 @@ use tokio_stream::wrappers::ReceiverStream;
 use tracing::{error, span, warn, Level};
 
 use super::{
-	indexer::rules::{
-		seed::{no_hidden, no_os_protected},
-		IndexerRule, RuleKind,
-	},
+	indexer::rules::seed::{no_hidden, no_os_protected},
 	normalize_path,
 };
