+++ crates/sd-indexer/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "sd-indexer"
+version = "0.0.1"
+license.workspace = true
+edition.workspace = true
+repository.workspace = true
+publish = false
+
+[dependencies]
+sd-utils = { path = "../utils" }
+
+chrono.workspace = true
+futures-util = "0.3.30"
+globset = { version = "0.4.14", features = ["serde1"] }
+opendal = "0.45.1"
+serde = { workspace = true, features = ["derive"] }
+specta.workspace = true
+thiserror.workspace = true
+tracing.workspace = true
+rmp-serde = "1.1.2"
+
+# TODO: Remove these
+rspc.workspace = true
+tokio = { workspace = true, features = ["fs"] }
+sd-prisma = { path = "../prisma" }
+tempfile.workspace = true
