+++ crates/sd-indexer/src/lib.rs
@@ -1,0 +1,4 @@
+mod ephemeral;
+pub mod rules;
+
+pub use ephemeral::ephemeral;
