+++ crates/sd-indexer/src/rules.rs
@@ -1,0 +1,9 @@
+mod error;
+mod rule_kind;
+mod rule_per_kind;
+mod rules;
+
+pub use error::*;
+pub use rule_kind::*;
+pub use rule_per_kind::*;
+pub use rules::*;
